+++ Runtime/Catalog/DreamyAudioCatalog.cs
@@ -8,10 +8,12 @@ namespace Dreamy.Audio
     {
         [SerializeField] private string catalogId = "core";
         [SerializeField] private string generatedNamespace = "Dreamy.Audio.Generated";
+        [SerializeField] private List<AudioLibrary> libraries = new List<AudioLibrary>();
         [SerializeField] private List<AudioEventDefinition> events = new List<AudioEventDefinition>();
 
         public string CatalogId => catalogId;
         public string GeneratedNamespace => generatedNamespace;
+        public IReadOnlyList<AudioLibrary> Libraries => libraries;
         public IReadOnlyList<AudioEventDefinition> Events => events;
 
         public bool TryGetEvent(string eventKey, out AudioEventDefinition definition)
@@ -27,6 +29,33 @@ namespace Dreamy.Audio
             }
 
             definition = null;
+            return false;
+        }
+
+        public bool TryGetFile(string eventKey, out AudioFileObject file)
+        {
+            for (var i = 0; i < libraries.Count; i++)
+            {
+                var library = libraries[i];
+                if (library == null)
+                {
+                    continue;
+                }
+
+                if (library.TryGetSound(eventKey, out var sound))
+                {
+                    file = sound;
+                    return true;
+                }
+
+                if (library.TryGetMusic(eventKey, out var music))
+                {
+                    file = music;
+                    return true;
+                }
+            }
+
+            file = null;
             return false;
         }
     }